requirements.txt
Flask==2.3.3
cryptography
google-genai
gunicorn==21.2.0
